+++ lib/mtl/rails/view_helpers.rb
@@ -226,17 +226,19 @@ module Mtl
       # @param menu [String, Boolean] (HTML ID) references the id of the aside menu / default
       #   nav menu to show / hide on mobile devices. The default used is `nav-menu`.
       #   When set to `false`, this button is skipped.
+      # @param class [String, Array] (HTML CLASS) additional, custom css class on header
       # @yield Additional content to be rendered as part of the header
       # @return [String] HTML safe string
-      def mtl_header(title = translate('.title', default: 'Menu'),
-                     back: false, menu: 'nav-menu', &block)
+      def mtl_header(title = translate('.title', default: 'Menu'), **options, &block)
         mtl_content = block_given? ? capture(&block) : nil
+        mtl_class = ['mtl-layout-default-header', options[:class]].compact.flatten.join(' ')
 
         render file: 'mtl/header', locals: {
           mtl_content: mtl_content,
           mtl_title: title.presence,
-          mtl_back: back,
-          mtl_menu: menu
+          mtl_back: options.fetch(:back, false),
+          mtl_menu: options.fetch(:menu, 'nav-menu'),
+          mtl_class: mtl_class
         }
       end